Refactor: Split Header component into smaller components
This commit refactors the Header component into smaller, more manageable components for improved maintainability. It breaks down the original component into a main Header component and separate components for mobile and desktop navigation elements.
